api/requirements.txt
pandas==2.1.4
numpy==1.26.2
aspose-pdf==23.5.0
pypdf2==3.0.1
openpyxl==3.1.2
xlsxwriter==3.1.9